backend/requirements.txt
pytz==2020.1  # https://github.com/stub42/pytz
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
redis==3.5.3  # https://github.com/andymccurdy/redis-py
#hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0  # https://github.com/celery/celery
#flower==0.9.5  # https://github.com/mher/flower
requests==2.23.0
lxml==4.5.1
psutil==5.6.5
beautifulsoup4==4.9.1
paramiko==2.7.1
ipip-ipdb==1.3.2
PyMySQL==0.9.3
python-ldap==3.3.1
itsdangerous==0.24

# TELEGRAM-BOT
# ------------------------------------------------------------------------------
python-telegram-bot==12.8

# Django
# ------------------------------------------------------------------------------
dwebsocket
django==2.1.4
#dwebsocket==0.5.12
django-auth-ldap==2.1.0 # https://github.com/django-auth-ldap/django-auth-ldap
django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
django-environ==0.4.5  # https://github.com/joke2k/django-environ
#django-redis==4.12.1  # https://github.com/jazzband/django-redis
gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
djangorestframework==3.11.0
django-filter==2.2.0
django-celery-results==1.2.1
# django-cors-headers==3.2.1

# Django
# ------------------------------------------------------------------------------
#django-storages[boto3]==1.9.1  # https://github.com/jschneier/django-storages
#django-anymail[mailgun]==7.2  # https://github.com/anymail/django-anymail

# Custom
pyvsphere
pyesxi